requirements.txt
groq
python-dotenv==1.0.0